+++ lib/src/document_actions_button.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:rapido/documents.dart';
+import 'package:rapido/rapido.dart';
 
 /// A button that will add editing options UI to a documentList document.
 /// It creates a builder button with a contextial menu, that allows editing