+++ lib/src/document_list_sort_button.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:rapido/documents.dart';
+import 'package:rapido/rapido.dart';
 
 /// A button that adds sorting to a DocumentList. Useful for adding
 /// sorting functionality to a DocumentListView.